Refactor API directory
