+++ src/lib/shared/api/adapters.ts
