+++ src/lib/shared/api/supabase.ts
@@ -1,5 +1,5 @@
 import { createClient } from '@supabase/supabase-js';
-import constants from '../constants';
+import constants from './credentials';
 
 const supabaseClient = createClient(constants.SUPABASE_URL, constants.SUPABASE_USER_KEY, {
   schema: 'public',
